requirements.txt
pymongo>=3.10.1<4.0.0
praw>=7.0.0,<8.0.0
python-twitter>=3.5,<4.0.0
gsheets>=0.4.1,<1.0.0
python-dotenv>=0.13.0,<1.0.0
requests>=2.22.0,<3.0.0
typing>=3.6.4,<4.0.0